Add files via upload
a very basic editable html page for chinese romanization but with no javascript.  What would be nice would be a way of saving changes (could be done by local storage, but would be better if connected to a spreadsheet/database).   Then multiple people could edit it.  Also versioning therefore would be nice.